post release backend error handling (#5306)
we were getting a lot of these errors in sentry
https://tldraw.sentry.io/issues/6152025342/?environment=production&environment=production-tldraw-multiplayer&project=4503966963662848&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&statsPeriod=14d&stream_index=1
which seemed to mostly the registerUser failing ?

But I couldn't figure out why, so I fixed the error reporting for that
method, and added a bunch of debugs to add breadcrumb context on sentry.
 
At the same time, I added a bunch of extra fail-safes to restart things
if the user durable objects are failing to connect to the replicator.
hopefully that will mitigate whatever is going on. this was causing the
durable object's in-memory cache to get stale and make it look like the
db writes were failing when in fact they were not.

### Change type

- [x] `other`

+++ apps/dotcom/sync-worker/src/UserDataSyncer.ts
@@ -102,7 +102,7 @@ export class UserDataSyncer {
 	replicator: TLPostgresReplicator
 
 	store = new OptimisticAppStore()
-	mutations: { mutationNumber: number; mutationId: string }[] = []
+	mutations: { mutationNumber: number; mutationId: string; timestamp: number }[] = []
 
 	sentry
 	private captureException(exception: unknown, extras?: Record<string, unknown>) {
@@ -117,8 +117,10 @@ export class UserDataSyncer {
 		}
 	}
 
+	interval
+
 	constructor(
-		ctx: DurableObjectState,
+		private ctx: DurableObjectState,
 		env: Environment,
 		private db: Kysely<DB>,
 		private userId: string,
@@ -129,11 +131,25 @@ export class UserDataSyncer {
 		this.sentry = createSentry(ctx, env)
 		this.replicator = getReplicator(env)
 		this.reboot(false)
+
+		this.interval = setInterval(() => this.__rebootIfMutationsNotCommitted(), 1000)
+	}
+
+	close() {
+		clearInterval(this.interval)
 	}
 
 	private queue = new ExecutionQueue()
 
+	numConsecutiveReboots = 0
+
 	async reboot(delay = true) {
+		this.numConsecutiveReboots++
+		if (this.numConsecutiveReboots > 5) {
+			this.logEvent({ type: 'user_do_abort', id: this.userId })
+			this.ctx.abort()
+			return
+		}
 		this.log.debug('rebooting')
 		this.logEvent({ type: 'reboot', id: this.userId })
 		await this.queue.push(async () => {
@@ -142,6 +158,7 @@ export class UserDataSyncer {
 			}
 			try {
 				await this.boot()
+				this.numConsecutiveReboots = 0
 			} catch (e) {
 				this.logEvent({ type: 'reboot_error', id: this.userId })
 				this.captureException(e)
@@ -357,6 +374,17 @@ export class UserDataSyncer {
 	async waitUntilConnected() {
 		while (this.state.type !== 'connected') {
 			await this.state.promise
+		}
+	}
+
+	private __rebootIfMutationsNotCommitted() {
+		// if any mutations have been not been committed for 5 seconds, let's reboot the cache
+		for (const mutation of this.mutations) {
+			if (Date.now() - mutation.timestamp > 5000) {
+				this.log.debug("Mutations haven't been committed for 5 seconds, rebooting")
+				this.reboot()
+				break
+			}
 		}
 	}
 }
